requirements.txt
asgiref==3.3.1
attrs==21.2.0
awsebcli==3.20.1
bcrypt==3.2.0
blessed==1.18.1
botocore==1.21.31
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-allauth==0.44.0
django-phonenumber-field==5.0.0
django-user-agents==0.4.0
djangorestframework==3.12.4
dnspython==1.16.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
httpie==2.4.0
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
oauthlib==3.1.0
paramiko==2.7.2
pathspec==0.5.9
phonenumberslite==8.12.19
Pillow==8.1.1
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pymongo==3.12.0
PyNaCl==1.4.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.4
ua-parser==0.10.0
urllib3==1.25.11
user-agents==2.2.0
wcwidth==0.1.9
websocket-client==0.59.0
